MSVC: Fix integer conversion warnings in containers
Add some casts, fixing warnings like:
src/corelib/text/qbytearray.h(490): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
src/corelib/text/qstring.h(1045): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
src/corelib/tools/qarraydatapointer.h(80): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
src/corelib/tools/qarraydatapointer.h(75): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
src/corelib/text/qbytearray.h(490): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

+++ src/corelib/text/qbytearray.cpp
@@ -1218,7 +1218,7 @@ QByteArray &QByteArray::operator=(const char *str)
         d = QByteArrayData(pair.first, pair.second, 0);
     } else {
         const int len = int(strlen(str));
-        const size_t fullLen = len + 1;
+        const uint fullLen = uint(len) + 1;
         if (d->needsDetach() || fullLen > d->allocatedCapacity()
                 || (len < size() && fullLen < (d->allocatedCapacity() >> 1)))
             reallocData(fullLen, d->detachFlags());
